// frontend/src/api/generated/models/ProtocolWorkflow.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ProtocolStep } from './ProtocolStep';
/**
 * Complete protocol workflow definition.
 */
export type ProtocolWorkflow = {
    workflow_id: string;
    workflow_name: string;
    steps?: Array<ProtocolStep>;
};