// frontend/src/api/generated/models/ProtocolFewShotExample.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Few-shot example in protocol format.
 */
export type ProtocolFewShotExample = {
    content: string;
    label: ProtocolFewShotExample.label;
    description: string;
};
export namespace ProtocolFewShotExample {
    export enum label {
        PASS = 'PASS',
        FAIL = 'FAIL',
    }
}